Moving the code to the base class as it is OAuth2 relevant code, not OIDC relevant code
